requirements.txt
paho-mqtt==2.1.0
python-dotenv==1.0.1
joblib==1.4.2
numpy==2.0.1
pandas==2.2.2
scikit-learn==1.5.1
tqdm==4.66.5
imblearn==0.0
imbalanced-learn==0.12.3
matplotlib==3.9.0
seaborn==0.13.2
scipy==1.14.0
keras==3.7.0